rhetenor/pyproject.toml
[project]
name = "rhetenor"
version = "0.1.2"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.3.5",
    "ollama>=0.6.0",
    "pydantic>=2.11.7",
    "butterflow",
    "morpho",
    "boto3>=1.40.18",
    "pytest>=8.4.2",
    "zstandard>=0.23.0",
    "PyYAML>=6.0.2"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rhetenor"]

[tool.uv.sources]
morpho = { git = "https://github.com/Thessal/transpiler" }
butterflow = { git = "https://github.com/Thessal/butterflow" }
